Create 1047-Remove-all-adjacent-duplicates-in-string
Solution: The solution of this problem requires an empty string concept. Take an empty string ans, and now iterate on the given string s and check in the ans string if the right-most character of the ans string matches the current character of the string s. If the current character matches the right-most character of ans string then pop_back() else push_back(s[i).